requirements.txt
flask==3.1.0
aiogram==3.18.0
#opencv-python==4.11.0.86
pillow==11.1.0
pymupdf==1.25.3
pdfplumber==0.11.5
python-docx==1.1.2
requests==2.32.3
numpy==2.2.3
torch==2.6.0
transformers
#openai-whisper==20231117
elevenlabs==1.52.0
gTTS==2.5.4 # Альтернатива TTS
jinja2==3.1.5
markupsafe==3.0.2
werkzeug==3.1.3
click==8.1.8
setuptools==75.8.2
pydantic==2.10.6
tqdm==4.67.1
regex==2024.11.6
diffusers==0.32.2
soundfile==0.13.1
pydub==0.25.1
docker==7.0.0 # Версия Docker SDK
nltk
python-dotenv #Для загрузки переменных окружения из .env файла
